Task_2_FinGPT_Powered_Compliance_Agents/fingpt-compliance-agents/src/data_collection/audio_data_collector.py
#!/usr/bin/env python3
"""
Audio Data Collector for FinGPT Compliance Agents
Collects financial audio data for training and evaluation
"""

import os
import json
import yaml
import argparse
import logging
from pathlib import Path
from typing import Dict, List, Any, Optional
import requests
from bs4 import BeautifulSoup
import pandas as pd
from tqdm import tqdm

logger = logging.getLogger(__name__)

class AudioDataCollector:
    """Collects financial audio data from various sources"""
    
    def __init__(self, config_path: str = "configs/config.yaml"):
        with open(config_path, 'r') as f:
            self.config = yaml.safe_load(f)
        
        self.output_dir = Path(self.config['output']['data_path']) / 'raw' / 'audio'
        self.output_dir.mkdir(parents=True, exist_ok=True)
        
        # Setup logging
        logging.basicConfig(
            level=logging.INFO,
            format='%(asctime)s - %(name)s - %(levelname)s - %(message)s'
        )
    
    def collect_earnings_calls(self, companies: List[str], max_calls: int = 10) -> List[Dict]:
        """Collect earnings call transcripts and audio"""
        logger.info("Collecting earnings calls data")
        
        earnings_data = []
        
        # This is a placeholder - in practice, you would need to:
        # 1. Access earnings call databases (Seeking Alpha, etc.)
        # 2. Download audio files
        # 3. Transcribe using ASR
        # 4. Process and format data
        
        for company in companies:
            logger.info(f"Collecting data for {company}")
            
            # Placeholder data structure
            call_data = {
                'company': company,
                'call_type': 'earnings_call',
                'date': '2024-01-01',  # Would be actual date
                'transcript': f"Earnings call transcript for {company}...",
                'audio_url': f"https://example.com/audio/{company}.mp3",
                'duration': 3600,  # seconds
                'participants': ['CEO', 'CFO', 'Analyst'],
                'topics': ['revenue', 'growth', 'outlook']
            }
            
            earnings_data.append(call_data)
        
        # Save collected data
        output_file = self.output_dir / "earnings_calls.json"
        with open(output_file, 'w') as f:
            json.dump(earnings_data, f, indent=2)
        
        logger.info(f"Collected {len(earnings_data)} earnings calls")
        return earnings_data
    
    def collect_financial_podcasts(self, max_episodes: int = 20) -> List[Dict]:
        """Collect financial podcast episodes"""
        logger.info("Collecting financial podcasts data")
        
        podcast_data = []
        
        # Placeholder for podcast collection
        # In practice, you would scrape podcast feeds or use APIs
        
        podcasts = [
            "The Motley Fool",
            "Marketplace",
            "Planet Money",
            "Financial Times",
            "Bloomberg Radio"
        ]
        
        for podcast in podcasts:
            for i in range(max_episodes // len(podcasts)):
                episode_data = {
                    'podcast': podcast,
                    'episode_title': f"Episode {i+1}: Financial Analysis",
                    'date': '2024-01-01',
                    'transcript': f"Podcast transcript for {podcast} episode {i+1}...",
                    'audio_url': f"https://example.com/podcasts/{podcast}_{i+1}.mp3",
                    'duration': 1800,  # 30 minutes
                    'topics': ['market_analysis', 'investment_strategy', 'economic_outlook']
                }
                
                podcast_data.append(episode_data)
        
        # Save collected data
        output_file = self.output_dir / "financial_podcasts.json"
        with open(output_file, 'w') as f:
            json.dump(podcast_data, f, indent=2)
        
        logger.info(f"Collected {len(podcast_data)} podcast episodes")
        return podcast_data
    
    def collect_financial_news_audio(self, max_segments: int = 30) -> List[Dict]:
        """Collect financial news audio segments"""
        logger.info("Collecting financial news audio data")
        
        news_data = []
        
        # Placeholder for news audio collection
        news_sources = [
            "CNBC",
            "Bloomberg",
            "Reuters",
            "Financial Times",
            "Wall Street Journal"
        ]
        
        for source in news_sources:
            for i in range(max_segments // len(news_sources)):
                segment_data = {
                    'source': source,
                    'title': f"Financial News Segment {i+1}",
                    'date': '2024-01-01',
                    'transcript': f"News segment transcript from {source}...",
                    'audio_url': f"https://example.com/news/{source}_{i+1}.mp3",
                    'duration': 300,  # 5 minutes
                    'topics': ['market_update', 'company_news', 'economic_indicators']
                }
                
                news_data.append(segment_data)
        
        # Save collected data
        output_file = self.output_dir / "financial_news.json"
        with open(output_file, 'w') as f:
            json.dump(news_data, f, indent=2)
        
        logger.info(f"Collected {len(news_data)} news segments")
        return news_data
    
    def create_training_data(self, audio_data: List[Dict]) -> List[Dict]:
        """Create training data from collected audio"""
        logger.info("Creating training data from audio")
        
        training_data = []
        
        for data in audio_data:
            # Create training examples for different tasks
            
            # 1. Sentiment analysis
            sentiment_example = {
                'instruction': 'Analyze the sentiment of this financial audio transcript.',
                'input': f"Transcript: {data['transcript']}",
                'output': 'Sentiment: neutral',  # Would be determined by analysis
                'task': 'audio_sentiment_analysis',
                'source': data.get('company', data.get('podcast', data.get('source', 'unknown')))
            }
            training_data.append(sentiment_example)
            
            # 2. Key information extraction
            extraction_example = {
                'instruction': 'Extract key financial information from this audio transcript.',
                'input': f"Transcript: {data['transcript']}",
                'output': 'Key information: Revenue, growth, outlook',  # Would be extracted
                'task': 'audio_information_extraction',
                'source': data.get('company', data.get('podcast', data.get('source', 'unknown')))
            }
            training_data.append(extraction_example)
            
            # 3. Topic classification
            topic_example = {
                'instruction': 'Classify the main topics discussed in this financial audio.',
                'input': f"Transcript: {data['transcript']}",
                'output': f"Topics: {', '.join(data.get('topics', []))}",
                'task': 'audio_topic_classification',
                'source': data.get('company', data.get('podcast', data.get('source', 'unknown')))
            }
            training_data.append(topic_example)
        
        # Save training data
        output_file = self.output_dir / "audio_training_data.jsonl"
        with open(output_file, 'w', encoding='utf-8') as f:
            for example in training_data:
                f.write(json.dumps(example, ensure_ascii=False) + '\n')
        
        logger.info(f"Created {len(training_data)} audio training examples")
        return training_data
    
    def collect_all_audio_data(self) -> Dict[str, List[Dict]]:
        """Collect all types of audio data"""
        logger.info("Starting comprehensive audio data collection")
        
        all_data = {}
        
        # Collect different types of audio data
        companies = ['Apple', 'Microsoft', 'Google', 'Amazon', 'Tesla']
        
        try:
            all_data['earnings_calls'] = self.collect_earnings_calls(companies)
        except Exception as e:
            logger.error(f"Error collecting earnings calls: {e}")
            all_data['earnings_calls'] = []
        
        try:
            all_data['podcasts'] = self.collect_financial_podcasts()
        except Exception as e:
            logger.error(f"Error collecting podcasts: {e}")
            all_data['podcasts'] = []
        
        try:
            all_data['news'] = self.collect_financial_news_audio()
        except Exception as e:
            logger.error(f"Error collecting news audio: {e}")
            all_data['news'] = []
        
        # Combine all audio data
        combined_data = []
        for data_type, data_list in all_data.items():
            combined_data.extend(data_list)
        
        # Create training data
        training_data = self.create_training_data(combined_data)
        all_data['training'] = training_data
        
        # Save summary
        summary = {
            'total_audio_segments': len(combined_data),
            'training_examples': len(training_data),
            'data_types': list(all_data.keys()),
            'output_directory': str(self.output_dir)
        }
        
        with open(self.output_dir / "collection_summary.json", 'w') as f:
            json.dump(summary, f, indent=2)
        
        logger.info(f"Audio data collection completed: {summary}")
        return all_data

def main():
    parser = argparse.ArgumentParser(description='Collect financial audio data')
    parser.add_argument('--config', default='configs/config.yaml', help='Config file path')
    
    args = parser.parse_args()
    
    collector = AudioDataCollector(args.config)
    collector.collect_all_audio_data()

if __name__ == "__main__":
    main()